Organized file structure into folders, updating imports as required
