+++ src/Components/DarkModeParent.tsx
@@ -1,5 +1,6 @@
-export let currentState = false;
 //Parent class to keep toggle state between pages
+export let currentState = false;
+
 export function updateCurrentState(){
     currentState = !currentState;
 }
