+++ src/Pages/SimpleReport.tsx
@@ -1,9 +1,9 @@
 import React, { useState } from 'react';
-import './Home.css';
-import './General.css';
-import './Report.css';
-import { DarkModeToggle, bodyClassName } from './DarkModeToggle';
-import LinkButton from './LinkButton';
+import '../Formatting/Home.css';
+import '../Formatting/General.css';
+import '../Formatting/Report.css';
+import { DarkModeToggle, bodyClassName } from '../Components/DarkModeToggle';
+import LinkButton from '../Components/LinkButton';
 import 'bootstrap/dist/css/bootstrap.min.css';
 import { Button, Form } from 'react-bootstrap';
