+++ src/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import './index.css';
-import RoutePaths from './RoutePaths';
+import RoutePaths from './Components/RoutePaths';
 import reportWebVitals from './reportWebVitals';
 import { HashRouter } from 'react-router-dom';
